% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowgraph.R
\name{dataflow_graph}
\alias{dataflow_graph}
\title{Computes and saves a dataflow graph}
\usage{
dataflow_graph(file = NULL)
}
\arguments{
\item{file}{Character, default \code{NULL}. File path for saving the dataflow graph.
The graph is not saved by default.}
}
\description{
An internal function that computes the dataflow graph. It
can be used in a knitr chunk to save the graph.
}

