% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowgraph.R
\name{plot.dep_graph}
\alias{plot.dep_graph}
\title{Plot knitr dataflow graph}
\usage{
\method{plot}{dep_graph}(x, y = "all", plot = TRUE, units = "by_chunk",
  ...)
}
\arguments{
\item{x}{Object of class \code{dep_graph}. A dependency graph for a knitr input file.}

\item{y}{Character, default \code{'all'}. Possible values are \code{'all'},
\code{'manual'} and \code{'auto'}, which selects which edges type are plotted.}

\item{plot}{Logical, default \code{TRUE}. Should the generated plot actually be plotted.}

\item{units}{Character, default \code{'by_chunk'}. Appropriate units on times and file sizes
are computed individually for each chunk by default. Set to \code{'all_same'} to
get the same unit across all chunks.}

\item{...}{Additional arguments passed on to \code{grViz}.}
}
\value{
A character string (invisibly) containing the graph in the dot format.
}
\description{
Visualization of the dependencies among chunks in a knitr input file.
Returns also the graph in the textual dot format (for Graphviz) that can processed
using the Graphviz command line tools.
}

