% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hooks.R
\name{set_hooks}
\alias{set_hooks}
\title{Sets knitr hook functions used by knitrflow}
\usage{
set_hooks()
}
\description{
When called, this function sets the hook functions \code{.__flow} and
\code{.__grab}. It is called when the
knitrflow package is attached, thus these hooks are set whenever the package
is loaded.
}
\details{
It can be used in a knitr document to set the hooks without loading knitrflow,
e.g. to make sure that the hook functions are set no matter how
the document is knitted.
}

